local_backup/requirements.txt
# GuardianShield Agents - Enhanced Dependencies
# Core dependencies for autonomous agent operations with recursive self-improvement

# Web3 and blockchain
web3>=6.0.0
eth-account>=0.9.0
eth-utils>=2.0.0

# Machine Learning and Analytics (for self-improvement)
scikit-learn>=1.3.0
numpy>=1.24.0
matplotlib>=3.7.0
pandas>=1.5.0

# FastAPI for API server and WebSocket support
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
websockets>=11.0.0

# HTTP requests and async
requests>=2.31.0
httpx>=0.24.0
aiohttp>=3.8.0

# Natural Language Processing (for threat analysis)
spacy>=3.6.0
beautifulsoup4>=4.12.0

# Security and encryption (enhances self-improvement safety)
cryptography>=41.0.0
python-dotenv>=1.0.0
pycryptodome>=3.18.0

# Genetic algorithm dependencies
regex>=2023.6.0

# Email notifications
secure-smtplib>=0.1.1

# Configuration and data formats
pyyaml>=6.0.1
toml>=0.10.2

# Enhanced logging and monitoring
structlog>=23.1.0

# Database support (for persistent storage)
sqlalchemy>=2.0.0

# Data validation
pydantic>=2.0.0

# Additional ML/AI capabilities (optional but recommended)
torch>=2.0.0; python_version>="3.8"
transformers>=4.30.0; python_version>="3.8"

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Performance monitoring
psutil>=5.9.0

# Network analysis
dnspython>=2.4.0